Create Role permission user , and access in role sidebar when login user in admin side

+++ app/Http/Controllers/Admin/RoleUserController.php
@@ -114,6 +114,9 @@ class RoleUserController extends Controller
     public function destroy(string $id)
     {
         $user = Admin::findOrFail($id);
+        if($user->getRoleNames()->first() === 'Administrator') {
+            return response(['status' => 'error','message'=>__('Deleted Successfully')]);
+        }
         $user->delete();
 
         return response(['status'=>'success','message'=> __('Deleted Succesffully')]);
